Fix more wn64 warnings.

+++ test/test-eof.c
@@ -78,7 +78,7 @@ main(int argc, char **argv)
 		return (1);
 
 
-	send(pair[0], test, strlen(test)+1, 0);
+	send(pair[0], test, (int)strlen(test)+1, 0);
 	shutdown(pair[0], SHUT_WR);
 
 	/* Initalize the event library */
